Change SaveGame mechanic
Made it so that the game only saves when changing realm from one to another and when claiming unlocks.

Previously game saved when changing mouse sensitivity, claiming unlocks, or crucially picking up crystals.

With the new implementation, progress is lost when perishing in a realm, e.g. drowning or fully exhausted